requirements.txt

# Install Python dependencies in stages
# First install numpy and basic scientific libraries
numpy==1.26.4

# Install core AWS and PDF libraries
boto3==1.38.0 
botocore==1.38.0 
PyPDF2==3.0.1 
pypdfium2==4.30.1 
pypdf==5.5.0 
pdfplumber==0.11.7

# Install image processing libraries
# removed opencv-python-headless==4.8.1.78 \
Pillow==11.3.0 #11.2.1 
opencv-python==4.11.0.86 
pdf2image==1.17.0

# Install pandas and data processing libraries
pandas==1.5.3 
tabula-py==2.10.0 
tabulate==0.9.0

# Install remaining dependencies
multiprocess>=0.70.14 
typing-extensions>=4.0.0 
langdetect==1.0.9 
tenacity==9.1.2 
python-dotenv 
amazon-textract-textractor==1.9.2 
amazon-textract-caller==0.2.4 
amazon-textract-response-parser==1.0.3 
pycryptodome==3.19.1 #3.15.0 
natsort==8.4.0 
litellm==1.63.14

